+++ app/src/main/java/interactivemapuniandes/model/data/RouteDTO.kt
@@ -1,0 +1,11 @@
+package interactivemapuniandes.model.data
+
+data class RouteDTO (
+    val hasUpcomingClass: Boolean,
+    val nextClass: NextClass,
+)
+
+data class NextClass(
+    val id: String,
+    val title: String,
+)
